+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+- map.update
+
+### Fixed
+- iterator.consume and iterator.count now propagate errors correctly.
 
 ## [0.3.0] - 2020.12.06